sidebar-service finished
